Fix off-by-one error in resetting offsets.
The offset should be one more than the largest offset found in the filenames
since the offset should be set to the next record to be read. Also get rid of
many +1/-1 adjustments to offset, offsets, and beginOffsets in
TopicPartitionWriter to minimize the number of places we have to adjust offset
information by 1.