Ensure the connector always has its own copy of the partition assignment.
This addresses two problems. First, it guarantees that even if the set returned
by context.assignment() or in the onPartitionsAssigned/onPartitionsRevoked are
modified, the task still has the correct assignment information. Second, the set
returned by the context.assignment() method is an unmodifiable set and the
existing code was trying to modify it on rebalances. Copying it avoids this
problem.