Fix rebalancing to simply discard temp files that are still in progress in onPartitionsRevoked.
More generally, this changes TopicPartitionWriter to discard in progress work
when it closes. While we could, in some cases, possibly save the temp file and
continue to use it if we are assigned the same partition in
onPartitionsAssigned, this is tricky enough to get right that for now it makes
more sense to stick with the simple and clearly correct version.