Removing retry logic from HiveMetaStore to fix the metastore connection bloat.
The doWithRetry method in HiveMetaStore.java blindly calls
client.reconnect() on TException. This is a very generic exception class
and includes acceptable and frequently occuring errors like
AlreadyExistsException. Since we use the RetryingMetaStoreClient which
already  implements a sane retrying logic (reconnect is only called
after socket timeout), the same logic in doWithRetry is unnecessary.

Renamed doWithRetry to doAction and added a comment on why retries at
this layer are not required.