CC-3950: Change the HDFS sink task to return to Connect offsets for records committed to HDFS
Modified the task’s `preCommit(…)` method to always return the offsets for only those records that have been committed to HDFS. Connect will then use the consumer to commit those offsets to the `__consumer_offsets` topic, which allows for simple monitoring of the approximate lag of the connector.

First, this change does not affect the HDFS connector’s exactly once behavior, because the connector still writes and uses the offsets it records in the files in HDFS; as long as these files are found in HDFS for the topic partitions, the HDFS connector will use these to determine where Connect should begin consuming. If the HDFS connector does not find the files in HDFS, it will let Connect and the consumer determine where to start, which will be based upon the committed offsets if found or upon the `consumer.auto.offset.reset` value specified in the worker config (which for Connect defaults to `earliest`) if no offsets are found. None of this behavior has changed.

What has changed is that, starting with CP 4.0.1, the HDFS sink connector never returned offsets, which means Connect never updating the offsets in `__consumer_offsets` for the topic partitions the connector read. Users could still change these offsets, and the connector would use these if no offsets were found in HDFS, but it never updated these offsets.

Now, the HDFS connector does return the offsets of the records from the various topic partitions that were most recently committed to HDFS, and Connect will thus periodically (based upon `consumer.offset.flush.interval.ms` in the worker config) write these offsets to the `__consumer_offsets` topic using the appropriate consumer group for the connector.

Because the offsets are committed to `__consumer_offsets` asynchronously, it is likely that the offsets in `__consumer_offsets` will be somewhat behind where the connector actually is. In fact, the offsets in `__consumer_offsets` should never be ahead of the records that the connector has written to HDFS. This allows the offsets as a simple way to estimate the upper limit of the actual connector lag.